tambah extract docx

+++ mppl/mppl/Control/ControlCek.cs
@@ -9,12 +9,13 @@ using iTextSharp.text.pdf.parser;
 using iTextSharp.text.pdf;
 using iTextSharp.text;
 using System.Text;
-using Bytescout.Document;
+//using Bytescout.Document;
 using mppl.Entitas;
 using System.Security;
 using System.Security.Cryptography;
 using System.Text;
-using mppl.mppl;
+using ICSharpCode.SharpZipLib.Zip;
+using System.Xml;
 
 namespace mppl.Control
 {
@@ -23,22 +24,25 @@ namespace mppl.Control
         List<int> finger;
         String teks;
         HttpServerUtility server;
-        private int _count;
-        public int count
-        {
-            get
-            {
-                return _count;
-            }
-        }
+        private const string ContentTypeNamespace =
+        @"http://schemas.openxmlformats.org/package/2006/content-types";
 
-        public delegate void EachDokumenChecked(object sender);
-        public event EachDokumenChecked passedOne;
+        private const string WordprocessingMlNamespace =
+            @"http://schemas.openxmlformats.org/wordprocessingml/2006/main";
+
+        private const string DocumentXmlXPath =
+            "/t:Types/t:Override[@ContentType=\"" +
+            "application/vnd.openxmlformats-officedocument." +
+            "wordprocessingml.document.main+xml\"]";
+
+        private const string BodyXPath = "/w:document/w:body";
+
+        private Stream docxFile;
+        private string docxFileLocation = "";
         
         public ControlCek()
         {
             finger = new List<int>();
-            teks = "";
         }
 
         public ControlCek(HttpServerUtility server)
@@ -58,27 +62,21 @@ namespace mppl.Control
                         Stream coba = input.FileContent;
                         if (input.PostedFile.ContentType == "application/pdf")
                             ekstrakPdf(coba);
-                        else
-                            ekstrakDoc(coba);
                         finger = Winnowing.getFingerprint(teks);
                         finger.Sort();
 
                         ModelDokumen dokumens = new ModelDokumen();
                         //generate filepath
-                        string fullpath, namafile;
-                        do
+                        MD5 md5 = MD5.Create();
+                        byte[] dataNamaFile = md5.ComputeHash(Encoding.UTF8.GetBytes(judul));
+                        StringBuilder sbNamaFile = new StringBuilder();
+                        for (int i = 0; i < dataNamaFile.Length; i++)
                         {
-                            MD5 md5 = MD5.Create();
-                            byte[] dataNamaFile = md5.ComputeHash(Encoding.UTF8.GetBytes(judul));
-                            StringBuilder sbNamaFile = new StringBuilder();
-                            for (int i = 0; i < dataNamaFile.Length; i++)
-                            {
-                                sbNamaFile.Append(dataNamaFile[i].ToString("x2"));
-                            }
-                            namafile = sbNamaFile.ToString();
+                            sbNamaFile.Append(dataNamaFile[i].ToString("x2"));
+                        }
+                        string namafile = sbNamaFile.ToString();
 
-                            fullpath = server.MapPath("~/fingerprint_dokumen/" + namafile);
-                        } while (System.IO.File.Exists(@fullpath));
+                        string fullpath = server.MapPath("~/fingerprint_dokumen/" + namafile);
 
                         //proses menulis ke file mulai dari sini (buat San)
                         //
@@ -87,13 +85,17 @@ namespace mppl.Control
                         //ControlCek tulisCek = new ControlCek();
                         //System.IO.File.WriteAllLines(/*@"direktoriProject"*/, tulisCek);
                         //System.IO.File.WriteAllLines(/*@"fingerprint_dokumen"*/, tulisCek);
-                        System.IO.FileStream fs = System.IO.File.Create(@fullpath);
-                        fs.Close();
-                        using (System.IO.StreamWriter sw = new System.IO.StreamWriter(@fullpath, true))
+
+                        if (!System.IO.File.Exists(@fullpath))
                         {
-                            for (int i = 0; i < finger.Count; i++)
+                            System.IO.FileStream fs = System.IO.File.Create(@fullpath);
+                            fs.Close();
+                            using (System.IO.StreamWriter sw = new System.IO.StreamWriter(@fullpath, true))
                             {
-                                sw.WriteLine(finger[i]);
+                                for (int i = 0; i < finger.Count; i++)
+                                {
+                                    sw.WriteLine(finger[i]);
+                                }
                             }
                         }
                         //edited by rian
@@ -115,13 +117,13 @@ namespace mppl.Control
             return false;
         }
 
-        public Dictionary<dokumen,double> cek(FileUpload input,double threshold)
+        public Dictionary<dokumen,double> cek(FileUpload input)
         {
             try
             {
-                if (input.PostedFile.ContentType == "application/msword" || input.PostedFile.ContentType == "application/pdf")
+                if (input.PostedFile.ContentType == "application/msword" || input.PostedFile.ContentType == "application/pdf" || input.PostedFile.ContentType == "application/vnd.openxmlformats-officedocument.wordprocessingml.document" )
                 {
-                    if (input.PostedFile.ContentLength < 10240000)
+                    if (input.PostedFile.ContentLength < 4096000)
                     {
                         Dictionary<dokumen, double> hasil;//berisi dokumen dokumen yang mirip
                         hasil = new Dictionary<dokumen, double>();
@@ -129,32 +131,30 @@ namespace mppl.Control
                         Stream coba = input.FileContent;
                         if (input.PostedFile.ContentType == "application/pdf")
                             ekstrakPdf(coba);
-                        else
-                            ekstrakDoc(coba);
-                       
-                        teks = teks.Trim();
+                        else if (input.PostedFile.ContentType == "application/vnd.openxmlformats-officedocument.wordprocessingml.document")
+                            ekstrakDocx(coba);
                         finger = Winnowing.getFingerprint(teks);
                         finger.Sort();
+                        //queri fingerprint dari db
                         ModelDokumen dokumens = new ModelDokumen();
                         List<dokumen> db = dokumens.get().ToList<dokumen>();
                         //panggil fungsi cekKemiripan dengan parameter list fingerprint dari file yang diupload sama yang didb
-                        _count = 0;
                         foreach (var i in db)
                         {
                             string[] read = File.ReadAllLines(server.MapPath("~/fingerprint_dokumen/" + i.alamat_fingerprint));
                             List<int> fingerprintdb = Array.ConvertAll<string, int>(read, new Converter<string, int>(Convert.ToInt32)).ToList<int>();
-                            double result = cekKemiripan(finger, fingerprintdb);
-                            if (result > threshold)
+                            double result;
+                            if(finger.Count>fingerprintdb.Count)
+                                result = cekKemiripan(fingerprintdb,finger);
+                            else
+                                result = cekKemiripan(finger, fingerprintdb);
+                            if (result > 0.5)
                             {
                                 //hasil fungsi masukan ke list
-                                hasil.Add(i, result);
-                                //hasil[i] = result;
+                                hasil[i] = result;
                             }
-                            _count++;
-                            //passedOne(this);
                         }
-                         
-                       // var result = cekKemiripan(finger, finger2);
+
                         //sorting list trus kembalikan List tersebut sebagai return value
                         
                         return hasil;
@@ -171,8 +171,7 @@ namespace mppl.Control
             }
             catch (Exception ex)
             {
-               // Console.WriteLine("file gagal diupload karena : " + ex);
-                Console.WriteLine(ex.StackTrace);
+                Console.WriteLine("file gagal diupload karena : " + ex);
                 return null;
             }
         }
@@ -183,24 +182,20 @@ namespace mppl.Control
             for (int i = 1; i <= reader.NumberOfPages; i++)
                 teks += PdfTextExtractor.GetTextFromPage(reader, i);
         }
-        void ekstrakDoc(Stream path)
+        void ekstrakDocx(Stream path)
         {
-            using (Bytescout.Document.Document doc = new Bytescout.Document.Document())
-            {
-                doc.Open(path);
-                StringBuilder sb = new StringBuilder();
-                for (int i = 0; i < doc.ParagraphCount; i++)
-                {
-                    Bytescout.Document.Paragraph p = doc.GetParagraph(i);
-                    if (p != null)
-                    {
-                        sb.Append(p.ToString());
-                        sb.AppendLine();
-                    }
-                }
+            docxFile = path;
+            //if (string.IsNullOrEmpty(docxFile))
+            //    throw new Exception("Input file not specified.");
 
-                teks = sb.ToString();
-            }
+            // Usually it is "/word/document.xml"
+
+            docxFileLocation = FindDocumentXmlLocation();
+
+            if (string.IsNullOrEmpty(docxFileLocation))
+                throw new Exception("It is not a valid Docx file.");
+
+            teks = ReadDocumentXml();
         }
         void createFingerPrint(string input)
         {
@@ -227,5 +222,144 @@ namespace mppl.Control
             }
             return counterIntersect / input.Count;
         }
+        #region FindDocumentXmlLocation()
+        /// 
+        /// Gets location of the "document.xml" zip entry.
+        /// 
+        /// Location of the "document.xml".
+        private string FindDocumentXmlLocation()
+        {
+            ZipFile zip = new  ZipFile(docxFile);
+            foreach (ZipEntry entry in zip)
+            {
+                // Find "[Content_Types].xml" zip entry
+
+                if (string.Compare(entry.Name, "[Content_Types].xml", true) == 0)
+                {
+                    Stream contentTypes = zip.GetInputStream(entry);
+
+                    XmlDocument xmlDoc = new XmlDocument();
+                    xmlDoc.PreserveWhitespace = true;
+                    xmlDoc.Load(contentTypes);
+                    contentTypes.Close();
+
+                    //Create an XmlNamespaceManager for resolving namespaces
+
+                    XmlNamespaceManager nsmgr =
+                        new XmlNamespaceManager(xmlDoc.NameTable);
+                    nsmgr.AddNamespace("t", ContentTypeNamespace);
+
+                    // Find location of "document.xml"
+
+                    XmlNode node = xmlDoc.DocumentElement.SelectSingleNode(
+                        DocumentXmlXPath, nsmgr);
+
+                    if (node != null)
+                    {
+                        string location =
+                            ((XmlElement)node).GetAttribute("PartName");
+                        return location.TrimStart(new char[] { '/' });
+                    }
+                    break;
+                }
+            }
+            zip.Close();
+            return null;
+        }
+        #endregion
+
+        #region ReadDocumentXml()
+        /// 
+        /// Reads "document.xml" zip entry.
+        /// 
+        /// Text containing in the document.
+        private string ReadDocumentXml()
+        {
+            StringBuilder sb = new StringBuilder();
+
+            ZipFile zip = new ZipFile(docxFile);
+            foreach (ZipEntry entry in zip)
+            {
+                if (string.Compare(entry.Name, docxFileLocation, true) == 0)
+                {
+                    Stream documentXml = zip.GetInputStream(entry);
+
+                    XmlDocument xmlDoc = new XmlDocument();
+                    xmlDoc.PreserveWhitespace = true;
+                    xmlDoc.Load(documentXml);
+                    documentXml.Close();
+
+                    XmlNamespaceManager nsmgr =
+                        new XmlNamespaceManager(xmlDoc.NameTable);
+                    nsmgr.AddNamespace("w", WordprocessingMlNamespace);
+
+                    XmlNode node =
+                        xmlDoc.DocumentElement.SelectSingleNode(BodyXPath, nsmgr);
+
+                    if (node == null)
+                        return string.Empty;
+
+                    sb.Append(ReadNode(node));
+
+                    break;
+                }
+            }
+            zip.Close();
+            return sb.ToString();
+        }
+        #endregion
+
+        #region ReadNode()
+        /// 
+        /// Reads content of the node and its nested childs.
+        /// 
+        /// XmlNode.
+        /// Text containing in the node.
+        private string ReadNode(XmlNode node)
+        {
+            if (node == null || node.NodeType != XmlNodeType.Element)
+                return string.Empty;
+
+            StringBuilder sb = new StringBuilder();
+            foreach (XmlNode child in node.ChildNodes)
+            {
+                if (child.NodeType != XmlNodeType.Element) continue;
+
+                switch (child.LocalName)
+                {
+                    case "t":                           // Text
+                        sb.Append(child.InnerText.TrimEnd());
+
+                        string space =
+                            ((XmlElement)child).GetAttribute("xml:space");
+                        if (!string.IsNullOrEmpty(space) &&
+                            space == "preserve")
+                            sb.Append(' ');
+
+                        break;
+
+                    case "cr":                          // Carriage return
+                    case "br":                          // Page break
+                        sb.Append(Environment.NewLine);
+                        break;
+
+                    case "tab":                         // Tab
+                        sb.Append("\t");
+                        break;
+
+                    case "p":                           // Paragraph
+                        sb.Append(ReadNode(child));
+                        sb.Append(Environment.NewLine);
+                        sb.Append(Environment.NewLine);
+                        break;
+
+                    default:
+                        sb.Append(ReadNode(child));
+                        break;
+                }
+            }
+            return sb.ToString();
+        }
+        #endregion
     }
 }
